Ultrasonic Calibration data
During development, we tested the ultrasonic sensor to work out the elevation of the sensor from the base of the sensor. 
The sensor has a wooden platform below it to stabilise the ultrasonic sensor. 
Data was collected by measuring the height from the base of the sensor platform to the bottom of the mass spring used, and plotting this value against the value given by the ultrasonic sensor. Finding a regression line, the y-intercept gave the offset.